🚚 move : 순한참조 해결 위한 SearchService에 getSearchProduct 메서드 분리
