+++ src/main/java/com/sparta/hotitemcollector/domain/product/ProductController.java
@@ -25,6 +25,7 @@ import org.springframework.web.bind.annotation.RestController;
 public class ProductController {
 
     private final ProductService productService;
+    private final SearchService searchService;
 
     @PostMapping
     public ResponseEntity<CommonResponse<ProductResponseDto>> createProduct(
@@ -97,11 +98,11 @@ public class ProductController {
 
     @GetMapping("/search")
     public ResponseEntity<CommonResponse<List<ProductSimpleResponseDto>>> getSearchProduct(
-        @RequestParam String nickname,
-        @RequestParam String productName,
-        @RequestParam ProductCategory category,
+        @RequestParam(required=false) String nickname,
+        @RequestParam(required=false) String productName,
+        @RequestParam(required=false) ProductCategory category,
         @RequestParam(defaultValue = "1") int page, @RequestParam(defaultValue = "10") int size) {
-        List<ProductSimpleResponseDto> responseDtoList = productService.getSearchProduct(
+        List<ProductSimpleResponseDto> responseDtoList = searchService.getSearchProduct(
             nickname,productName,category,page-1,size);
         CommonResponse<List<ProductSimpleResponseDto>> response = new CommonResponse<>(
             "검색을 통한 상품 목록 조회 성공", 200, responseDtoList);
